Ejercicio7

+++ Ejercicio7.java
@@ -10,7 +10,7 @@ public class Ejercicio7 {
     public static void main (String args []){
         Scanner sc = new Scanner(System.in);
         System.out.print("Dime el total de la base imponible de la factura: ");
-        double n = sc.nextInt();
+        double n = sc.nextDouble();
 
         
         System.out.print("El total de la factura es: " + n+(n*0.21));
